fix: reconstruct dirs
