+++ src/features/appoints/appointsIndex.ts
@@ -1,7 +1,7 @@
-import { fetchAppoints } from "../../../apis/appoint";
-import { Appoint } from "../../../utils/type";
-import { updateAppoints } from "../../reducers/appoints";
-import store from "../../store";
+import { fetchAppoints } from "../../apis/appoint";
+import { Appoint } from "../../utils/type";
+import { updateAppoints } from "../../redux/reducers/appoints";
+import store from "../../redux/store";
 
 export const requestAppoints = async (startStr: string, endStr: string) => {
   const appoints: Appoint[] = await fetchAppoints(startStr, endStr).then(
