+++ src/features/auth/signup.ts
@@ -1,8 +1,8 @@
 import Router from "next/router";
-import { fetchLogin } from "../../../apis/auth";
-import { getAccessToken, signUpWithGoogle } from "../../../utils/auth";
-import { loginSuccess } from "../../reducers/auth";
-import store from "../../store";
+import { fetchLogin } from "../../apis/auth";
+import { loginSuccess } from "../../redux/reducers/auth";
+import store from "../../redux/store";
+import { getAccessToken, signUpWithGoogle } from "../../utils/auth";
 
 // actionをこのように扱うのは違うかな。redux-toolkitのDocを読もう https://redux-toolkit.js.org/tutorials/typescript#define-slice-state-and-action-types
 export const requestRegister = () => {
