+++ src/pages/auth/index.tsx
@@ -1,8 +1,32 @@
-import React from "react";
-import Auth from "../../components/auth";
+import { useDispatch } from "react-redux";
+import { fetchSessionCheck } from "../../apis/auth";
+import { requestLogin } from "../../features/auth/login";
+import { requestRegister } from "../../features/auth/signup";
 
-const index = () => {
-  return <Auth></Auth>;
+const Auth = () => {
+  const dispatch = useDispatch();
+  const handleRegister = () => {
+    dispatch(requestRegister());
+  };
+
+  const handleLogin = () => {
+    dispatch(requestLogin());
+  };
+
+  const handleCurrentUser = async () => {
+    await fetchSessionCheck().then((data) => data.currentUser);
+  };
+  return (
+    <>
+      <div>ユーザー登録</div>
+      <button onClick={handleRegister}>Googleでユーザー登録</button>
+      <br />
+      <div>ログイン</div>
+      <button onClick={handleLogin}>Googleでログイン</button>
+      <br />
+      <button onClick={handleCurrentUser}>ログイン中のユーザー</button>
+    </>
+  );
 };
 
-export default index;
+export default Auth;
