+++ src/pages/index.tsx
@@ -3,7 +3,7 @@ import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { fetchSessionCheck } from "../apis/auth";
 import Calender from "../components/calender";
-import { requestLogout } from "../redux/actions/auth/logout";
+import { requestLogout } from "../features/auth/logout";
 import { isLoggedIn } from "../utils/auth";
 
 export default function Home() {
