+++ src/utils/firebase/signup.ts
@@ -2,6 +2,5 @@ import axios from "axios";
 import { signup } from "../../config/urls";
 
 export const fetchSignUp = async (uid: string, name: string) => {
-  //TODO: このaxiosはインスタンス化して使いまわしたい https://github.com/axios/axios#creating-an-instance
   return await axios.post(signup, { user: { uid, name } }).then((res) => res);
 };
